[C] Track `snd-bpe` when the available window is too small to send the next available data fragment.

+++ aeron-client/src/main/c/concurrent/aeron_term_scanner.c
@@ -16,5 +16,5 @@
 
 #include "concurrent/aeron_term_scanner.h"
 
-extern size_t aeron_term_scanner_scan_for_availability(
-    const uint8_t *buffer, size_t term_length_left, size_t max_length, size_t *padding);
+extern int32_t aeron_term_scanner_scan_for_availability(
+    const uint8_t *buffer, int32_t term_length_left, int32_t max_length, int32_t *padding);
